backend/requirements.txt
Flask == 2.0.1
Flask-Cors == 3.0.10
python-dotenv == 0.21.1
requests == 2.28.2 
Flask-API == 3.0.post1
polyline == 2.0.0
folium == 0.14.0